Add indicator of message sending
This is totally a hack that doesn't really indicate the message went
anywhere. It makes senders feel better that pressing the button did
something.

If I could figure out how to get a status out of pusher, I'd do
something different. Putting the sent into the greeting push function
results in the recipient getting the status update, not the sender.